moved User.ts file
