+++ game/src/gameCore/race/ServerRaceGameController.ts
@@ -18,8 +18,8 @@ import PlayerState from "../../communication/race/PlayerState";
 import RaceGameState from "../../communication/race/RaceGameState";
 import UserInfo from "../../communication/user/UserInfo";
 import QuestionRepository from "../../server/data/QuestionRepository";
-import User from "../../server/data/User";
 import Room from "../../server/rooms/Room";
+import User from "../../server/rooms/User";
 import { getObjectValues } from "../../utils/Utils";
 import { Clock } from "../clock/Clock";
 import { ServerGame } from "../Game";
