+++ game/src/server/rooms/User.ts
@@ -1,6 +1,7 @@
 import UserInfo from "../../communication/user/UserInfo";
 
 export default interface User {
+	isReady: boolean;
 	userId: string;
 	userInfo: UserInfo;
 	socket: SocketIO.Socket;
